hubble/recorder: Fix race in queue closing
When the queue is closed in s.close(), we set it to nil to avoid
accidental send-on-closed-channel panics. In the sink go routine
however, we were fetching the reference to s.queue without holding the
lock, thus potentially introducing a race.

This commit addresses this issue by copying the reference to the channel
before we start reading from it.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>