hubble/recorder: Improve ergonomics by not closing stats channel
Since we have introduced the `Handle.Done` channel, we do not have toV
signal the shutdown of the sink by closing the statistics channel
anymore. Instead, consumers can now wait on the `Handle.Done` channel
getting closed.

While there are not many benefits in this version of the code, it will
make the select statement in a subsequent commit much more readable.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>