hubble/recorder: Implement stop conditions
This implements a feature to stop a recording as soon as one of the
three stop conditions are hit. We push the conditions into the sink, to
ensure they are hit as early as possible, e.g. to avoid writing more
bytes or packets than necessary.

While this can also be implemented client-side by observing the
statistics of the running recording and issuing an explicit stop
request, it can be useful to implement some basic stop conditions
server-side as a fail-safe in case the client becomes unresponsive but
is still connected.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>