hubble/recorder: Stop without draining queue
Previously, we waited for the recording queue to drain when the client
requested a stop. However, because the client has no visibility into the
queue (and indeed doesn't even know if there are queued records when
they issue a stop request), this does not provide any value to the
client.

Therefore, this PR changes the semantics of a stop request by
immediately initiating a shutdown, instead of waiting for the queue to
drain. This ensures that the resulting recording more closely matches
the observed statistics at the time when the client issued a stop
request. It also simplifies the code a bit.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>