ipcache: Release IPCache write lock during UpdatePolicyMaps
This ensures that the endpoint code can read IPCache while the update is
in progress. If we hold on to the IPCache lock, the endpoint code will
deadlock.

To allow releasing the IPCache write lock while performing a
label-triggered identity update, we introduce a new 'updating' mutex
which must always be held before the IPCache write lock is acquired. By
holding on to the 'updating' lock in removeLabelsFromIPs, we ensure that
we will be able to re-acquire the write lock once all reads have
finished.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>