clusterpool-v2beta: Refactor CIDR release and status
This commit refactors the internals of the clusterpool-v2beta
implementation a bit, to prepare the code for clusterpool-v2beta2, which
will share some of the code. We intend to deprecate v2beta in favor of
v2beta2 in a future release, but for now there is some code sharing
involved.

The first change in this commit is adding a method to the nodeUpdater
interface. It's not used by the old implementation, but will be used in
a subsequent commit.

The other change is changing the behavior of the `status()` method,
which the new code will implement in a different manner. To accommodate
for that, the allocation and release thresholds are moved out of the
pool definition itself, and are instead passed whenever the new status
(and thus new allocations and releases) are to be computed.

The status method (now called clusterPoolV2Beta1Status) now also
performs CIDR release in a separate helper function, instead of inline
while generating the status. This simplifies the code further.

Co-authored-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>