agent: Add clusterpool-v2beta2 allocator
This commit adds the agent side of the clusterpool-v2beta2 allocator.
Most of the logic is shared with the existing clusterpool-v2beta
allocator. The main differences are that v2beta2 supports multiple
pools, and explicitly announces the number of needed IPs to the
operator.  That number of needed IPs is computed as the number of
allocated IPs, rounded up the next multiple of a per-pool pre-allocate
value. This value will also be used to determine if the node has excess
CIDRs, in which case the node will attempt to release any unused CIDRs
back into the IPAM pool by removing it from its CiliumNode CRD.

The pools that the agent will request from the operator are statically
configured in via the newly added hidden `cluster-pool-node-pre-alloc`
flag. This flag is intended to be temporary. The IPAM pools a node
requires will instead be derived from the networks the node is attached
to. That feature to support multi-homed nodes will be added in a
subsequent PR.

Co-authored-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Tobias Klauser <tobias@cilium.io>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>